tcdiff/src/MICA/requirements_OLD.txt
absl-py==1.1.0
aiohttp==3.8.4
aiosignal==1.3.1
albumentations==1.2.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==22.2.0
Bottleneck @ file:///tmp/build/80754af9/bottleneck_1648028895253/work
brotlipy==0.7.0
cachetools==5.2.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///opt/conda/conda-bld/cffi_1642701102775/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
chumpy==0.70
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1598884132938/work
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1602866480661/work
coloredlogs==15.0.1
contourpy==1.0.7
cryptography @ file:///tmp/build/80754af9/cryptography_1639400846433/work
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask @ file:///tmp/build/80754af9/dask-core_1602083700509/work
decorator==4.4.2
easydict==1.9
faiss==1.7.3
flatbuffers==2.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.3.0
fvcore==0.1.5.post20210915
gast==0.4.0
google-auth==2.9.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
graphviz==0.8.4
grpcio==1.47.0
h5py==3.8.0
humanfriendly==10.0
idna @ file:///tmp/build/80754af9/idna_1637925883363/work
imageio @ file:///tmp/build/80754af9/imageio_1594161405741/work
importlib-metadata==4.12.0
importlib-resources==5.11.0
insightface==0.6.2
iopath==0.1.9
joblib==1.1.0
keras==2.11.0
kiwisolver==1.2.0
libclang==15.0.6.1
lightning-utilities==0.8.0
loguru==0.6.0
Markdown==3.3.7
matplotlib==3.7.0
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626186064646/work
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
mxnet==1.9.1
networkx @ file:///tmp/build/80754af9/networkx_1598376031484/work
numexpr @ file:///tmp/build/80754af9/numexpr_1640704208950/work
numpy==1.24.2
oauthlib==3.2.0
onnx==1.12.0
onnxruntime==1.11.0
onnxruntime-gpu==1.13.1
opencv-python==4.7.0.68
opencv-python-headless==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas==1.4.1
Pillow==9.4.0
plot-model==0.20
portalocker @ file:///home/conda/feedstock_root/build_artifacts/portalocker_1645465631507/work
prettytable==3.3.0
protobuf==3.19.4
provider==0.1.0
psbody-mesh==0.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyOpenGL==3.1.6
pyOpenSSL @ file:///opt/conda/conda-bld/pyopenssl_1643788558760/work
pyparsing==2.4.7
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytorch-ignite==0.4.11
pytorch-lightning==1.9.4
pytorch3d==0.7.1
pytz==2021.3
PyWavelets @ file:///tmp/build/80754af9/pywavelets_1601658317819/work
PyYAML==6.0
pyzmq==25.0.0
qudida==0.0.4
requests @ file:///opt/conda/conda-bld/requests_1641824580448/work
requests-oauthlib==1.3.1
rsa==4.8
scikit-image==0.17.2
scikit-learn==1.1.1
scipy==1.10.0
seaborn==0.12.2
six @ file:///tmp/build/80754af9/six_1644875935023/work
sympy==1.11.1
tabulate @ file:///home/conda/feedstock_root/build_artifacts/tabulate_1614001031686/work
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2020.10.1
toolz @ file:///tmp/build/80754af9/toolz_1601054250827/work
torch==1.11.0
torchaudio==0.11.0
torchmetrics==0.11.4
torchvision==0.12.0
tornado @ file:///tmp/build/80754af9/tornado_1606942300299/work
tqdm @ file:///home/conda/feedstock_root/build_artifacts/tqdm_1648096753605/work
trimesh==3.12.7
tsnecuda==3.0.1+cu112
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
UNKNOWN==0.0.0
urllib3 @ file:///opt/conda/conda-bld/urllib3_1643638302206/work
wcwidth==0.2.5
Werkzeug==1.0.1
wrapt==1.15.0
yacs @ file:///home/conda/feedstock_root/build_artifacts/yacs_1645705974477/work
yarl==1.8.2
zipp==3.8.0
